ServiceRecipeRepository id wrong type Long (vs
ServiceRecipeRepository id wrong type Long (vs Integer).

Issue-ID: SO-2329

+++ mso-catalog-db/src/main/java/org/onap/so/db/catalog/data/repository/ServiceRecipeRepository.java
@@ -25,7 +25,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "serviceRecipe", path = "serviceRecipe")
-public interface ServiceRecipeRepository extends JpaRepository<ServiceRecipe, Long> {
+public interface ServiceRecipeRepository extends JpaRepository<ServiceRecipe, Integer> {
     ServiceRecipe findByActionAndServiceModelUUID(String action, String serviceModelUUID);
 
     ServiceRecipe findByAction(String action);
